+++ modules/AWSPowerShell/Cmdlets/EMRServerless/Basic/New-EMRServerlessApplication-Cmdlet.cs
@@ -103,6 +103,18 @@ namespace Amazon.PowerShell.Cmdlets.EMRServerless
         public System.Int32? AutoStopConfiguration_IdleTimeoutMinute { get; set; }
         #endregion
         
+        #region Parameter ImageConfiguration_ImageUri
+        /// <summary>
+        /// <para>
+        /// <para>The URI of an image in the Amazon ECR registry. This field is required when you create
+        /// a new application. If you leave this field blank in an update, Amazon EMR will remove
+        /// the image configuration.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.String ImageConfiguration_ImageUri { get; set; }
+        #endregion
+        
         #region Parameter InitialCapacity
         /// <summary>
         /// <para>
@@ -136,7 +148,7 @@ namespace Amazon.PowerShell.Cmdlets.EMRServerless
         #region Parameter ReleaseLabel
         /// <summary>
         /// <para>
-        /// <para>The EMR release version associated with the application.</para>
+        /// <para>The EMR release associated with the application.</para>
         /// </para>
         /// </summary>
         #if !MODULAR
@@ -198,6 +210,22 @@ namespace Amazon.PowerShell.Cmdlets.EMRServerless
         #endif
         [Amazon.PowerShell.Common.AWSRequiredParameter]
         public System.String Type { get; set; }
+        #endregion
+        
+        #region Parameter WorkerTypeSpecification
+        /// <summary>
+        /// <para>
+        /// <para>The key-value pairs that specify worker type to <code>WorkerTypeSpecificationInput</code>.
+        /// This parameter must contain all valid worker types for a Spark or Hive application.
+        /// Valid worker types include <code>Driver</code> and <code>Executor</code> for Spark
+        /// applications and <code>HiveDriver</code> and <code>TezTask</code> for Hive applications.
+        /// You can either set image details in this parameter for each worker type, or in <code>imageConfiguration</code>
+        /// for all worker types.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("WorkerTypeSpecifications")]
+        public System.Collections.Hashtable WorkerTypeSpecification { get; set; }
         #endregion
         
         #region Parameter ClientToken
@@ -278,6 +306,7 @@ namespace Amazon.PowerShell.Cmdlets.EMRServerless
             context.AutoStopConfiguration_Enabled = this.AutoStopConfiguration_Enabled;
             context.AutoStopConfiguration_IdleTimeoutMinute = this.AutoStopConfiguration_IdleTimeoutMinute;
             context.ClientToken = this.ClientToken;
+            context.ImageConfiguration_ImageUri = this.ImageConfiguration_ImageUri;
             if (this.InitialCapacity != null)
             {
                 context.InitialCapacity = new Dictionary<System.String, Amazon.EMRServerless.Model.InitialCapacityConfig>(StringComparer.Ordinal);
@@ -320,6 +349,14 @@ namespace Amazon.PowerShell.Cmdlets.EMRServerless
                 WriteWarning("You are passing $null as a value for parameter Type which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
             }
             #endif
+            if (this.WorkerTypeSpecification != null)
+            {
+                context.WorkerTypeSpecification = new Dictionary<System.String, Amazon.EMRServerless.Model.WorkerTypeSpecificationInput>(StringComparer.Ordinal);
+                foreach (var hashKey in this.WorkerTypeSpecification.Keys)
+                {
+                    context.WorkerTypeSpecification.Add((String)hashKey, (WorkerTypeSpecificationInput)(this.WorkerTypeSpecification[hashKey]));
+                }
+            }
             
             // allow further manipulation of loaded context prior to processing
             PostExecutionContextLoad(context);
@@ -391,6 +428,25 @@ namespace Amazon.PowerShell.Cmdlets.EMRServerless
             if (cmdletContext.ClientToken != null)
             {
                 request.ClientToken = cmdletContext.ClientToken;
+            }
+            
+             // populate ImageConfiguration
+            var requestImageConfigurationIsNull = true;
+            request.ImageConfiguration = new Amazon.EMRServerless.Model.ImageConfigurationInput();
+            System.String requestImageConfiguration_imageConfiguration_ImageUri = null;
+            if (cmdletContext.ImageConfiguration_ImageUri != null)
+            {
+                requestImageConfiguration_imageConfiguration_ImageUri = cmdletContext.ImageConfiguration_ImageUri;
+            }
+            if (requestImageConfiguration_imageConfiguration_ImageUri != null)
+            {
+                request.ImageConfiguration.ImageUri = requestImageConfiguration_imageConfiguration_ImageUri;
+                requestImageConfigurationIsNull = false;
+            }
+             // determine if request.ImageConfiguration should be set to null
+            if (requestImageConfigurationIsNull)
+            {
+                request.ImageConfiguration = null;
             }
             if (cmdletContext.InitialCapacity != null)
             {
@@ -480,6 +536,10 @@ namespace Amazon.PowerShell.Cmdlets.EMRServerless
             {
                 request.Type = cmdletContext.Type;
             }
+            if (cmdletContext.WorkerTypeSpecification != null)
+            {
+                request.WorkerTypeSpecifications = cmdletContext.WorkerTypeSpecification;
+            }
             
             CmdletOutput output;
             
@@ -546,6 +606,7 @@ namespace Amazon.PowerShell.Cmdlets.EMRServerless
             public System.Boolean? AutoStopConfiguration_Enabled { get; set; }
             public System.Int32? AutoStopConfiguration_IdleTimeoutMinute { get; set; }
             public System.String ClientToken { get; set; }
+            public System.String ImageConfiguration_ImageUri { get; set; }
             public Dictionary<System.String, Amazon.EMRServerless.Model.InitialCapacityConfig> InitialCapacity { get; set; }
             public System.String MaximumCapacity_Cpu { get; set; }
             public System.String MaximumCapacity_Disk { get; set; }
@@ -556,6 +617,7 @@ namespace Amazon.PowerShell.Cmdlets.EMRServerless
             public System.String ReleaseLabel { get; set; }
             public Dictionary<System.String, System.String> Tag { get; set; }
             public System.String Type { get; set; }
+            public Dictionary<System.String, Amazon.EMRServerless.Model.WorkerTypeSpecificationInput> WorkerTypeSpecification { get; set; }
             public System.Func<Amazon.EMRServerless.Model.CreateApplicationResponse, NewEMRServerlessApplicationCmdlet, object> Select { get; set; } =
                 (response, cmdlet) => response;
         }