+++ modules/AWSPowerShell/Cmdlets/Lightsail/Basic/New-LSDomain-Cmdlet.cs
@@ -50,10 +50,7 @@ namespace Amazon.PowerShell.Cmdlets.LS
         #region Parameter DomainName
         /// <summary>
         /// <para>
-        /// <para>The domain name to manage (e.g., <code>example.com</code>).</para><note><para>You cannot register a new domain name using Lightsail. You must register a domain
-        /// name using Amazon Route 53 or another domain name registrar. If you have already registered
-        /// your domain, you can enter its name in this parameter to manage the DNS records for
-        /// that domain using Lightsail.</para></note>
+        /// <para>The domain name to manage (e.g., <code>example.com</code>).</para>
         /// </para>
         /// </summary>
         #if !MODULAR