+++ modules/AWSPowerShell/Cmdlets/RDS/Basic/Edit-RDSDBInstance-Cmdlet.cs
@@ -145,7 +145,11 @@ namespace Amazon.PowerShell.Cmdlets.RDS
         #region Parameter CACertificateIdentifier
         /// <summary>
         /// <para>
-        /// <para>Specifies the certificate to associate with the DB instance.</para><para>This setting doesn't apply to RDS Custom.</para>
+        /// <para>Specifies the CA certificate identifier to use for the DB instance’s server certificate.</para><para>This setting doesn't apply to RDS Custom.</para><para>For more information, see <a href="https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html">Using
+        /// SSL/TLS to encrypt a connection to a DB instance</a> in the <i>Amazon RDS User Guide</i>
+        /// and <a href="https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html">
+        /// Using SSL/TLS to encrypt a connection to a DB cluster</a> in the <i>Amazon Aurora
+        /// User Guide</i>.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]