+++ modules/AWSPowerShell/Cmdlets/RDS/Basic/New-RDSDBInstance-Cmdlet.cs
@@ -121,6 +121,20 @@ namespace Amazon.PowerShell.Cmdlets.RDS
         public System.String BackupTarget { get; set; }
         #endregion
         
+        #region Parameter CACertificateIdentifier
+        /// <summary>
+        /// <para>
+        /// <para>Specifies the CA certificate identifier to use for the DB instance’s server certificate.</para><para>This setting doesn't apply to RDS Custom.</para><para>For more information, see <a href="https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html">Using
+        /// SSL/TLS to encrypt a connection to a DB instance</a> in the <i>Amazon RDS User Guide</i>
+        /// and <a href="https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html">
+        /// Using SSL/TLS to encrypt a connection to a DB cluster</a> in the <i>Amazon Aurora
+        /// User Guide</i>.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.String CACertificateIdentifier { get; set; }
+        #endregion
+        
         #region Parameter CharacterSetName
         /// <summary>
         /// <para>
@@ -860,6 +874,7 @@ namespace Amazon.PowerShell.Cmdlets.RDS
             context.AvailabilityZone = this.AvailabilityZone;
             context.BackupRetentionPeriod = this.BackupRetentionPeriod;
             context.BackupTarget = this.BackupTarget;
+            context.CACertificateIdentifier = this.CACertificateIdentifier;
             context.CharacterSetName = this.CharacterSetName;
             context.CopyTagsToSnapshot = this.CopyTagsToSnapshot;
             context.CustomIamInstanceProfile = this.CustomIamInstanceProfile;
@@ -977,6 +992,10 @@ namespace Amazon.PowerShell.Cmdlets.RDS
             if (cmdletContext.BackupTarget != null)
             {
                 request.BackupTarget = cmdletContext.BackupTarget;
+            }
+            if (cmdletContext.CACertificateIdentifier != null)
+            {
+                request.CACertificateIdentifier = cmdletContext.CACertificateIdentifier;
             }
             if (cmdletContext.CharacterSetName != null)
             {
@@ -1240,6 +1259,7 @@ namespace Amazon.PowerShell.Cmdlets.RDS
             public System.String AvailabilityZone { get; set; }
             public System.Int32? BackupRetentionPeriod { get; set; }
             public System.String BackupTarget { get; set; }
+            public System.String CACertificateIdentifier { get; set; }
             public System.String CharacterSetName { get; set; }
             public System.Boolean? CopyTagsToSnapshot { get; set; }
             public System.String CustomIamInstanceProfile { get; set; }